Cambios

+++ apps/documentos/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, redirect
 from django.conf import settings
 from .forms import ConsultaForm, DocumentoForm
-from .google_drive import search_files_in_drive, download_file, extract_text_from_file, get_drive_service  # Importar funciones de Google Drive
+from .google_drive import search_files_in_drive, download_file, extract_text_from_file, get_drive_service
 import logging
 import os
 from PIL import Image
@@ -11,21 +11,42 @@ import re
 import docx
 import openpyxl
 from datetime import datetime
-
+import hashlib
+from django.urls import reverse
 from django.http import HttpResponse, JsonResponse
+from .google_drive import get_drive_service, search_files_in_drive, vaciar_carpeta_drive, descargar_archivo
+from django.contrib.auth.decorators import user_passes_test
+from google.oauth2.credentials import Credentials
+from googleapiclient.discovery import build
+from googleapiclient.errors import HttpError
 
-from .google_drive import get_drive_service, search_files_in_drive
+# Configurar ruta al token
+TOKEN_PATH = os.path.join(settings.BASE_DIR, 'token.json')
 
-# Configura el logging
-logging.basicConfig(level=logging.INFO)
+# Configura el logger
 logger = logging.getLogger(__name__)
 
 pytesseract.pytesseract.tesseract_cmd = r'C:\Program Files\Tesseract-OCR\tesseract.exe'
 
-def extract_text_from_image(image_path):
+def preprocess_image(image_path):
     try:
         image = Image.open(image_path)
-        text = pytesseract.image_to_string(image)
+        # Convertir la imagen a escala de grises
+        image = image.convert('L')
+        # Aplicar filtros adicionales si es necesario
+        image = image.point(lambda x: 0 if x < 150 else 255, '1')  # Cambiar umbral según necesidad
+        return image
+    except Exception as e:
+        logger.error(f"Error al preprocesar la imagen {image_path}: {e}")
+        return None
+
+def extract_text_from_image(image_path):
+    try:
+        # Asegúrate de que las rutas sean válidas
+        image = Image.open(rf'{image_path}')
+        # Configuraciones de Tesseract
+        config = '--oem 3 --psm 6'
+        text = pytesseract.image_to_string(image, config=config, lang='spa')
         return text
     except Exception as e:
         logger.error(f"Error al extraer texto de la imagen {image_path}: {e}")
@@ -67,109 +88,81 @@ def extract_text_from_xlsx(xlsx_path):
         return ""
 
 def limpiar_texto(texto):
-    """Elimina caracteres no alfanuméricos de un texto y lo convierte a minúsculas."""
-    return re.sub(r'\W+', ' ', texto).lower()
+    """Elimina caracteres especiales y convierte el texto a minúsculas para una comparación más robusta."""
+    return re.sub(r'[^A-Za-z0-9\s]', '', texto).strip().lower()
 
 def buscar_termino(texto, consulta):
+    """Realiza la búsqueda del término en el texto. Considera palabras similares o cercanas."""
     texto_limpio = limpiar_texto(texto)
     consulta_limpia = limpiar_texto(consulta)
 
-    # Depuración: Mostrar la consulta y el texto procesados
-    print(f"Texto limpio: {texto_limpio[:1000]}")  # Solo mostrar los primeros 1000 caracteres
-    print(f"Consulta limpia: {consulta_limpia}")
+    if consulta_limpia in texto_limpio:
+        return True
 
-    return consulta_limpia in texto_limpio
+    texto_palabras = texto_limpio.split()
+    consulta_palabras = consulta_limpia.split()
 
+    for palabra in consulta_palabras:
+        if any(palabra in palabra_texto for palabra_texto in texto_palabras):
+            return True
+
+    return False
+
+# Vista de consulta con búsqueda solo en modo local
 def consulta_view(request):
+    resultados = None
+    cantidad_archivos = 0
+    search_done = False
+
     if request.method == 'POST':
         form = ConsultaForm(request.POST)
         if form.is_valid():
+            search_done = True
             consulta = form.cleaned_data['consulta']
-            origen = form.cleaned_data['origen']
-            logger.info(f"Consulta: {consulta}, Origen: {origen}")
-            print(f"Consulta: {consulta}, Origen: {origen}")
+            logger.info(f"Consulta: {consulta}")
             resultados = []
 
-            if origen == 'drive':
-                # Buscar archivos en la carpeta de Google Drive
-                drive_files = search_files_in_drive(settings.DRIVE_FOLDER_ID)
-                service = get_drive_service()
+            archivos_dir = settings.ARCHIVOS_DIR
+            for root, dirs, files in os.walk(archivos_dir):
+                for file_name in files:
+                    try:
+                        archivo_path = os.path.join(root, file_name)
+                        texto = ""
 
-                if not drive_files:
-                    logger.info("No se encontraron archivos en la carpeta de Google Drive.")
-                    print("No se encontraron archivos en la carpeta de Google Drive.")
-                else:
-                    for file in drive_files:
-                        try:
-                            file_id = file['id']
-                            file_name = file['name']
-                            mime_type = file['mimeType']
-                            created_time = file['createdTime']
-                            created_time_formatted = datetime.strptime(created_time, '%Y-%m-%dT%H:%M:%S.%fZ').strftime('%d/%m/%Y')
-                            web_view_link = file.get('webViewLink', '#')
-                            file_data = download_file(service, file_id)
-                            if file_data:
-                                texto = extract_text_from_file(file_data, mime_type)
-                                print(f"Texto extraído de {file_name}: {texto[:500]}")  # Mostrar los primeros 500 caracteres para depuración
-                                if buscar_termino(texto, consulta):
-                                    resultados.append({
-                                        'nombre': file_name,
-                                        'url': web_view_link,
-                                        'fecha': created_time_formatted
-                                    })
-                                    logger.info(f"Término encontrado en: {file_name}")
-                                    print(f"Término encontrado en: {file_name}")
-                                else:
-                                    logger.info(f"Término no encontrado en: {file_name}")
-                                    print(f"Término no encontrado en: {file_name}")
-                        except Exception as e:
-                            logger.error(f"Error procesando archivo {file_name}: {e}")
-                            print(f"Error procesando archivo {file_name}: {e}")
-            elif origen == 'local':
-                # Buscar archivos en la carpeta local
-                archivos_dir = settings.ARCHIVOS_DIR
-                for root, dirs, files in os.walk(archivos_dir):
-                    for file_name in files:
-                        try:
-                            archivo_path = os.path.join(root, file_name)
-                            mime_type = None
-                            texto = ""
+                        if file_name.lower().endswith('.pdf'):
+                            texto = extract_text_from_pdf(archivo_path)
+                        elif file_name.lower().endswith(('.png', '.jpg', '.jpeg')):
+                            texto = extract_text_from_image(archivo_path)
+                        elif file_name.lower().endswith('.docx'):
+                            texto = extract_text_from_docx(archivo_path)
+                        elif file_name.lower().endswith('.xlsx'):
+                            texto = extract_text_from_xlsx(archivo_path)
 
-                            if file_name.lower().endswith('.pdf'):
-                                mime_type = 'application/pdf'
-                                texto = extract_text_from_pdf(archivo_path)
-                            elif file_name.lower().endswith(('.png', '.jpg', '.jpeg')):
-                                mime_type = 'image/jpeg'
-                                texto = extract_text_from_image(archivo_path)
-                            elif file_name.lower().endswith('.docx'):
-                                mime_type = 'application/vnd.openxmlformats-officedocument.wordprocessingml.document'
-                                texto = extract_text_from_docx(archivo_path)
-                            elif file_name.lower().endswith('.xlsx'):
-                                mime_type = 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
-                                texto = extract_text_from_xlsx(archivo_path)
-
-                            if buscar_termino(texto, consulta):
+                        if buscar_termino(texto, consulta):
+                            # Verificar si el archivo está en la carpeta de documentos o descargados
+                            ruta_archivo = obtener_ruta_archivo(file_name)
+                            if ruta_archivo:
                                 fecha_modificacion = datetime.fromtimestamp(os.path.getmtime(archivo_path)).strftime('%d/%m/%Y')
                                 resultados.append({
                                     'nombre': file_name,
-                                    'url': f"{settings.MEDIA_URL}documentos/descargados/{file_name}",
+                                    'url': ruta_archivo,
                                     'fecha': fecha_modificacion
                                 })
-                                logger.info(f"Término encontrado en: {file_name}")
-                                print(f"Término encontrado en: {file_name}")
-                            else:
-                                logger.info(f"Término no encontrado en: {file_name}")
-                                print(f"Término no encontrado en: {file_name}")
-                        except Exception as e:
-                            logger.error(f"Error procesando archivo {file_name}: {e}")
-                            print(f"Error procesando archivo {file_name}: {e}")
+                    except Exception as e:
+                        logger.error(f"Error procesando archivo {file_name}: {e}")
 
-            return render(request, 'documentos/resultados.html', {'form': form, 'resultados': resultados, 'cantidad_archivos': len(resultados)})
+            cantidad_archivos = len(resultados)
 
     else:
         form = ConsultaForm()
 
-    return render(request, 'documentos/consulta.html', {'form': form})
+    context = {
+        'form': form,
+        'resultados': resultados if resultados else None,
+        'cantidad_archivos': cantidad_archivos,
+        'search_done': search_done
+    }
+    return render(request, 'documentos/consulta.html', context)
 
 def subir_comprobante_view(request):
     if request.method == 'POST':
@@ -181,6 +174,27 @@ def subir_comprobante_view(request):
         form = DocumentoForm()
     return render(request, 'documentos/subir_comprobante.html', {'form': form})
 
+def archivo_existe(ruta_descarga, nombre_archivo):
+    archivo_path = os.path.join(ruta_descarga, nombre_archivo)
+    return os.path.exists(archivo_path)
+
+# Verifica si el archivo ya existe en la ruta de descarga o en la carpeta principal de documentos.
+def obtener_ruta_archivo(file_name):
+    # Ruta en /media/documentos
+    ruta_documentos = os.path.join(settings.MEDIA_ROOT, 'documentos', file_name)
+    
+    # Ruta en /media/documentos/descargados
+    ruta_descargados = os.path.join(settings.MEDIA_ROOT, 'documentos', 'descargados', file_name)
+    
+    # Verificar si el archivo existe en alguna de las dos rutas
+    if os.path.exists(ruta_documentos):
+        return os.path.join(settings.MEDIA_URL, 'documentos', file_name)
+    elif os.path.exists(ruta_descargados):
+        return os.path.join(settings.MEDIA_URL, 'documentos', 'descargados', file_name)
+    else:
+        return None
+    
+
 # Nueva función: Descargar archivos desde Google Drive
 def descargar_archivos_nube(request):
     try:
@@ -189,13 +203,14 @@ def descargar_archivos_nube(request):
 
         if not drive_files:
             logger.info("No se encontraron archivos en la carpeta de Google Drive.")
-            return HttpResponse("No se encontraron archivos en la carpeta de Google Drive.")
-
-        logger.info(f"Archivos encontrados: {drive_files}")
+            return render(request, 'documentos/sin_archivos.html')
 
         ruta_descarga = os.path.join(settings.MEDIA_ROOT, 'documentos', 'descargados')
         if not os.path.exists(ruta_descarga):
             os.makedirs(ruta_descarga)
+
+        archivos_descargados = []
+        archivos_omitidos = []
 
         for file in drive_files:
             try:
@@ -204,29 +219,52 @@ def descargar_archivos_nube(request):
                 file_data = download_file(service, file_id)
 
                 if file_data:
-                    # Obtener la fecha de subida y formatearla
                     created_time = file.get('createdTime', '')
-                    fecha_formateada = datetime.strptime(created_time, "%Y-%m-%dT%H:%M:%S.%fZ").strftime("%Y-%m-%d_%H-%M-%S")
 
-                    # Renombrar el archivo con el formato deseado
-                    nuevo_nombre = f"bethania-{fecha_formateada}{os.path.splitext(file_name)[1]}"  # Mantiene la extensión original
+                    if created_time:
+                        try:
+                            fecha_formateada = datetime.strptime(created_time, "%Y-%m-%dT%H:%M:%S.%fZ").strftime("%Y%m%d_%H%M%S")
+                        except ValueError:
+                            logger.error(f"Error al formatear la fecha {created_time}. Asignando 'Fecha desconocida'.")
+                            fecha_formateada = "Fecha_desconocida"
+                    else:
+                        logger.warning(f"El archivo {file_name} no tiene 'createdTime'. Asignando 'Fecha desconocida'.")
+                        fecha_formateada = "Fecha_desconocida"
 
-                    archivo_path = os.path.join(ruta_descarga, nuevo_nombre)
-                    with open(archivo_path, 'wb') as archivo_local:  # 'wb' para escribir en modo binario
-                        archivo_local.write(file_data)  # Escribimos los bytes directamente
-                    logger.info(f"Archivo {file_name} descargado exitosamente como {nuevo_nombre}.")
+                    extension = os.path.splitext(file_name)[1]
+                    id_corto = file_id[:8]
+                    nuevo_nombre = f"bethania-{fecha_formateada}-{id_corto}{extension}"
+
+                    if archivo_existe(ruta_descarga, nuevo_nombre):
+                        logger.info(f"El archivo {nuevo_nombre} ya existe. Omitiendo descarga.")
+                        archivos_omitidos.append(nuevo_nombre)
+                    else:
+                        archivo_path = os.path.join(ruta_descarga, nuevo_nombre)
+                        with open(archivo_path, 'wb') as archivo_local:
+                            archivo_local.write(file_data)
+                        logger.info(f"Archivo {file_name} descargado exitosamente como {nuevo_nombre}.")
+                        archivos_descargados.append(nuevo_nombre)
+
                 else:
                     logger.error(f"No se pudo obtener el contenido del archivo {file_name}.")
             except Exception as e:
                 logger.error(f"Error descargando archivo {file_name}: {e}")
-                return HttpResponse(f"Error descargando archivo {file_name}: {e}")
+                return render(request, 'documentos/error_descarga.html', {'mensaje_error': f"Error descargando archivo {file_name}: {e}"})
 
-        return HttpResponse(f"Archivos descargados exitosamente en la carpeta {ruta_descarga}.")
+        return render(request, 'documentos/resumen_descarga.html', {
+            'archivos_descargados': archivos_descargados,
+            'archivos_omitidos': archivos_omitidos
+        })
     except Exception as e:
         logger.error(f"Error en la descarga de archivos: {e}")
-        return HttpResponse(f"Error en la descarga de archivos: {e}")
+        return render(request, 'documentos/error_descarga.html', {'mensaje_error': f"Error en la descarga de archivos: {e}"})
 
-# Nueva función: Vaciar la carpeta de Google Drive
+def es_superadministrador(user):
+    """Verifica si el usuario es un superadministrador."""
+    return user.is_superuser
+
+# Nueva función: Vaciar la carpeta de Google Drive (solo para superadministradores)
+@user_passes_test(es_superadministrador, login_url='/forbidden/')
 def vaciar_carpeta_drive(request):
     try:
         service = get_drive_service()
@@ -234,13 +272,20 @@ def vaciar_carpeta_drive(request):
         results = service.files().list(q=f"'{folder_id}' in parents").execute()
         files = results.get('files', [])
 
+        if not files:
+            return render(request, 'documentos/carpeta_vacia.html')
+
         for file in files:
             try:
                 service.files().delete(fileId=file['id']).execute()
             except Exception as error:
                 logger.error(f"Error al eliminar el archivo {file['name']}: {error}")
 
-        return JsonResponse({'mensaje': 'Carpeta vaciada correctamente.'})
+        return render(request, 'documentos/carpeta_vaciada_exito.html')
     except Exception as e:
         logger.error(f"Error vaciando la carpeta: {e}")
         return HttpResponse(f"Error vaciando la carpeta: {e}")
+
+def forbidden_view(request):
+    """Muestra un mensaje de error cuando el usuario no tiene permisos suficientes."""
+    return render(request, 'forbidden.html')
